refactor: extract database and models into modules
